requirements.txt
websocket-client
SimpleWebSocketServer